refactor: restructure project to use Cargo workspace
