+++ mc-booter/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "mc-booter"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tokio = { version = "1.46.1", features = ["full"] }
+tokio-util = "0.7.15"
+clap = { version = "4.5.41", features = ["derive"] }
+tracing = "0.1.41"
+num_cpus = "1.17.0"
