+++ mc-common/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "mc-common"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+regex = "1.11.1"
