+++ mc-register/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "mc-register"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.141"
+async-trait = "0.1.88"
+tracing = "0.1.41"
+reqwest = "0.12.22"
+toml = "0.9.4"
