+++ mc-register/src/external_api.rs
@@ -1,4 +1,4 @@
-use crate::service::register::{McpServer, Registry};
+use crate::{McpServer, Registry};
 use reqwest::Client;
 use serde::Deserialize;
 use std::error::Error;
