--- src/service/mod.rs
@@ -1,6 +1,0 @@
-pub mod config;
-pub mod proxy;
-pub mod register;
-mod router;
-pub mod server;
-mod session;
